Thanh nhan

+++ src/main/java/com/service/implement/FileLogService.java
@@ -8,8 +8,8 @@ public class FileLogService implements IFileLogService {
     private final IFileLogDao fileLogDao = new FileLogDao();
 
     @Override
-    public int insert(int configId, String path, String status, int authorId) {
-        return fileLogDao.insert(configId,path, status,  authorId);
+    public long insert(int configId, String path, String status, int authorId) {
+        return fileLogDao.insert(configId, path, status, authorId);
     }
 
     @Override
